pyproject.toml
[project]
name = "fall-in"
version = "0.1.0"
description = "헤쳐 모여! (Fall In!) - A strategic card game based on 6 Nimmt! with Korean Air Force theme"
readme = "README.md"
authors = [
    { name = "bnbong", email = "bbbong9@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "pygame-ce>=2.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
fall-in = "fall_in.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fall_in"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

